Fixing issues found during testing and prepare for public release
Setup a toggle for debugging. Clean the description of the menu item for Rebooting with a delay.  Fixed the output from the subprocess to return a string. Enabled Rebooting options. Added a note if Rebooting with a delay, it might get the outlet out of sync. Removed the PluginConfig.xml file, it was only used for enabling disabling debugging. Updated the README to include all the features that are enabled.